+++ src/main/java/net/mcreator/cravingsmod/procedures/GrapeSeedsRightclickedOnBlockProcedure.java
@@ -11,7 +11,7 @@ public class GrapeSeedsRightclickedOnBlockProcedure {
 	public static void execute(LevelAccessor world, double x, double y, double z, ItemStack itemstack) {
 		if ((world.getBlockState(BlockPos.containing(x, y, z))).getBlock() == Blocks.FARMLAND) {
 			world.setBlock(BlockPos.containing(x, y + 1, z), CravingsModModBlocks.GRAPE_CROP.get().defaultBlockState(), 3);
-			itemstack.setCount((int) (itemstack.getCount() - 1));
+			itemstack.setCount(itemstack.getCount() - 1);
 		}
 	}
 }
